pyproject.toml
[project]
name = "newsapitracker"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.6",
    "langchain-ollama>=1.0.1",
    "pandas>=2.1.0",
    "playwright>=1.57.0",
    "plotly>=5.18.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "schedule>=1.2.2",
    "sentencepiece>=0.2.1",
    "streamlit>=1.31.0",
    "tokenizers>=0.20.3",
    "torch>=2.0.0",
    "trafilatura>=2.0.0",
    "transformers>=4.57.6",
]
